refactor: branch cache
