+++ packages/storage/sqlite-node/src/sqlite/repo.ts
@@ -19,17 +19,10 @@ import {
 	type SessionStorage,
 } from "@earendil-works/pi-agent-core/experimental";
 import { type Usage, uuidv7 } from "@earendil-works/pi-ai";
+import { appendEntryToBranchCache, buildCachedBranch, deleteBranchCache, rebuildBranchCache } from "./branch-cache.ts";
 import { applyMigrations } from "./migrations.ts";
-import {
-	appendEntryToBranchCache,
-	buildCachedBranch,
-	type CachedBranchEntryRow,
-	deleteBranchCache,
-	queryCachedBranchRows,
-	readBranchTipIds,
-	readCachedBranch,
-	rebuildBranchCache,
-} from "./storage/branch-cache.ts";
+import { type CachedBranchEntryRow, queryCachedBranchRows, readCachedBranch } from "./storage/branch-entries.ts";
+import { readBranchTipIds } from "./storage/branch-tips.ts";
 import {
 	countMessageEntries,
 	deleteEntryRows,
